Feat(icons): Generate Icons as a React Component in Typescript
