+++ packages/icons/package.json
@@ -13,9 +13,19 @@
     "directory": "packages/icons"
   },
   "scripts": {
-    "build": "shx rm -rf dist && shx mkdir -p dist/svg && shx cp -r package.json README.md dist/ && node ./scripts/build.js"
+    "prepare:react": "node ./scripts/prepareSvgReact.js",
+    "build:svg": "node ./scripts/buildSvg.js",
+    "build:clean": "shx rm -rf dist",
+    "build:prepare": "shx mkdir -p dist/svg dist/react dist/.tmp-svg && shx cp -r package.json README.md dist/",
+    "build:react": "svg2react-icon --typescript --no-sub-dir ./dist/.tmp-svg ./dist/react",
+    "build:react:clean": "shx rm -rf dist/react/**.d.ts dist/react/**.js.map dist/.tmp-svg",
+    "prebuild": "npm-run-all build:clean build:prepare",
+    "build": "npm-run-all build:svg prepare:react build:react",
+    "postbuild": "yarn build:react:clean"
   },
   "devDependencies": {
-    "shx": "0.3.4"
+    "shx": "0.3.4",
+    "svg2react-icon": "^3.1.178",
+    "npm-run-all": "4.1.5"
   }
 }
